Update uberbot to use yaml config files
Rewrite of how we do command handling for uberbot.  In the past we had a
config file that was homegrown, then converted into a readable string that
could also be used to configure additional bots from the bot connection
itself.  This added parsing code and complexity that wasn't needed, and the
feature of starting new bots from existing bot connections was never used.

For better code cleanliness, that's all been pulled out and rewritten to a
more standardized data structure and parsing handlers.  At the same time,
there is now a command to update the old configuration file into a YAML
file, to bring old bots over to the new uberbot.

Closes #1